+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - chore: change host to new address ([#137](https://github.com/PostHog/posthog-android/pull/137))
+- fix: rename groupProperties to groups for capture methods ([#137](https://github.com/PostHog/posthog-android/pull/137))
 
 ## 3.3.0 - 2024-05-24